Added c++ library support

+++ mopsr/src/sigproc/read_header.c
@@ -134,6 +134,6 @@ int read_header (FILE *inputfile) /* includefile */
 
 
   /* return total number of bytes read */
-  return ftell(inputfile);
+  return (int) ftell(inputfile);
 }
